Added dired-dwim-target
https://emacs.stackexchange.com/questions/5603/how-to-quickly-copy-move-file-in-emacs-dired
dired-dwim-target is a variable defined in `dired.el'. Its value is nil

Documentation: If non-nil, Dired tries to guess a default target directory. This means: if there is a Dired buffer displayed in the next window, use its current directory, instead of this Dired buffer's current directory.

The target is used in the prompt for file copy, rename etc.